interference-agents/requirements.txt
claude-agent-sdk==0.1.52
python-dotenv>=1.0.0,<2.0.0
